Refactor CRSF Frame Handling and Remove Unused Classes
- Deleted CrsfFlightModeFrame and associated header files to streamline frame handling.
- Consolidated frame decoding logic into CrsfFrame class, enhancing maintainability.
- Updated frame data structures to use std::variant for better type safety and flexibility.
- Removed GpsPayload and related classes, integrating GPS data handling directly into CrsfFrame.
- Simplified Flight class to manage frames more efficiently, with dedicated vectors for each frame type.
- Enhanced ELRSMonitor to utilize the new frame structure and decoding methods.
- Updated tests to reflect changes in frame handling and removed references to deleted classes.

+++ src/Crsf/CrsfFrameData.hpp
@@ -2,8 +2,8 @@
 #include <variant>
 #include <string>
 #include <cstdint>
+#include <vector>
 
-// Forward declarations for frame data types
 struct GpsFrameData {
     double latitude;   // degrees
     double longitude;  // degrees
@@ -31,20 +31,17 @@ struct FlightModeFrameData {
 };
 
 struct LinkRXFrameData {
-    uint16_t rssi;      // dBm
-    uint16_t lq;        // Link Quality %
-    uint16_t noise;     // dBm
-    uint16_t txPower;   // dBm
+    int16_t rssi;       // RSSI in dBm (negative values like -50 to -100)
+    uint8_t lq;         // Link Quality percentage (0-100%)
+    int8_t noise;       // SNR in dB
+    uint8_t txPower;    // TX Power enum (0=0mW, 1=10mW, 2=25mW, 3=100mW, 4=500mW, 5=1000mW, 6=2000mW)
 };
-
-#include <vector>
 
 struct UnknownFrameData {
     uint8_t frameType;
     std::vector<uint8_t> rawPayload;
 };
 
-// Variant type that can hold any frame data
 using CrsfFrameData = std::variant<
     GpsFrameData,
     AttitudeFrameData,
